Do not backup production deployments.

+++ deploy.sh
@@ -92,7 +92,11 @@ tar czf - "$CIRCLE_PROJECT_REPONAME/" | eval $CMD_SSH_TAR
 DEPLOY_FILES_PATH="$CFG_FILES_PATH/$CIRCLE_PROJECT_REPONAME"
 CMD_SSH_FILES="cp -R $DEPLOY_FILES_PATH/* $TMPDIR/$CIRCLE_PROJECT_REPONAME"
 # Make a backup of the old deployment.
-CMD_BACKUP="if [ -d $DEPLOY_PATH ]; then rm -rf $DEPLOY_PATH.old; mv $DEPLOY_PATH $DEPLOY_PATH.old; fi"
+CMD_BACKUP=""
+if [ "$1" -ne "production" ]; then
+	CMD_BACKUP="if [ -d $DEPLOY_PATH ]; then rm -rf $DEPLOY_PATH.old; mv $DEPLOY_PATH $DEPLOY_PATH.old; fi"
+fi
+
 # Move the completed stream to the correct deploy path.
 CMD_MOVE_DEPLOYMENT="mkdir -p $DEPLOY_PATH; mv $TMPDIR/$CIRCLE_PROJECT_REPONAME/* $DEPLOY_PATH"
